@@ -1,14 +1,29 @@
 {
     "name": "ewallet/messaging",
     "description": "Ewallet messaging",
-    "license": "MIT",
     "version": "1.0",
+    "license": "MIT",
     "authors": [
         {
             "name": "Luis Montealegre",
             "email": "montealegreluis@gmail.com"
         }
     ],
+    "require": {
+        "ewallet/domain": "~1.0",
+        "symfony/console": "~4.2",
+        "vlucas/phpdotenv": "~2.0",
+        "zendframework/zend-mail": "~2.10"
+    },
+    "require-dev": {
+        "fzaninotto/faker": "~1.8",
+        "mockery/mockery": "~1.2",
+        "phpunit/phpunit": "~7.4",
+        "roave/security-advisories": "dev-master"
+    },
+    "config": {
+        "bin-dir": "bin/"
+    },
     "autoload": {
         "psr-4": {
             "Ewallet\\": "src/"
@@ -16,63 +31,13 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Ewallet\\": "vendor/ewallet/domain/tests/src",
-            "Hexagonal\\": [
-                "vendor/hexagonal/hexagonal/tests/src",
-                "vendor/hexagonal/messaging/tests/src"
-            ]
+            "Ewallet\\": "vendor/ewallet/domain/tests/src"
         }
-    },
-    "config": {
-        "bin-dir": "bin/"
     },
     "repositories": [
         {
             "type": "path",
             "url": "../../packages/ewallet/domain"
-        },
-        {
-            "type": "path",
-            "url": "../../packages/ewallet/application"
-        },
-        {
-            "type": "path",
-            "url": "../../packages/ewallet/doctrine"
-        },
-        {
-            "type": "path",
-            "url": "../../packages/ewallet/definitions"
-        },
-        {
-            "type": "path",
-            "url": "../../packages/ewallet/templating"
-        },
-        {
-            "type": "path",
-            "url": "../../packages/hexagonal/hexagonal"
-        },
-        {
-            "type": "path",
-            "url": "../../packages/hexagonal/doctrine"
-        },
-        {
-            "type": "path",
-            "url": "../../packages/hexagonal/messaging"
         }
-    ],
-    "require": {
-        "symfony/console": "^3.1",
-        "vlucas/phpdotenv": "^2.0",
-        "ewallet/application": "^1.0",
-        "ewallet/doctrine": "^1.0",
-        "ewallet/definitions": "^1.0",
-        "ewallet/templating": "^1.0",
-        "hexagonal/messaging": "^1.0",
-        "zendframework/zend-mail": "^2.5"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "^5.4",
-        "mockery/mockery": "^0.9",
-        "fzaninotto/faker": "^1.5"
-    }
+    ]
 }
