@@ -4,11 +4,12 @@
  *
  * This source file is subject to the license that is bundled with this package in the file LICENSE.
  */
+
 return [
     'doctrine' => [
         'mapping_dirs' => [
-            __DIR__ . '/vendor/ewallet/doctrine/src/Doctrine2/Resources/config',
-            __DIR__ . '/vendor/hexagonal/doctrine/src/Doctrine2/Resources/config',
+            __DIR__ . '/vendor/ewallet/domain/src/Ports/Doctrine/Ewallet/Resources/config',
+            __DIR__ . '/vendor/ewallet/domain/src/Ports/Doctrine/Application/Resources/config',
         ],
         'dev_mode' => getenv('APP_ENV') !== 'production',
         'proxy_dir' => __DIR__ . '/var/doctrine/proxies',
@@ -17,7 +18,7 @@ return [
             'path' => __DIR__ . '/var/ewallet.sq3',
         ],
         'types' => [
-            'MemberId' => 'Ewallet\Doctrine2\Types\MemberIdType',
+            'MemberId' => 'Ports\Doctrine\Ewallet\Types\MemberIdType',
         ],
     ],
     'twig' => [
