@@ -4,14 +4,13 @@
  *
  * This source file is subject to the license that is bundled with this package in the file LICENSE.
  */
+
 namespace Ewallet\Pimple;
 
-use Ewallet\Pimple\ServiceProviders\{
-    DoctrineServiceProvider,
-    EwalletMessagingServiceProvider,
-    HexagonalServiceProvider,
-    TwigServiceProvider
-};
+use Application\DependencyInjection\DoctrineServiceProvider;
+use Application\DependencyInjection\TwigServiceProvider;
+use Ewallet\Pimple\ServiceProviders\EwalletMessagingServiceProvider;
+use Ewallet\Pimple\ServiceProviders\MessagingServiceProvider;
 use Pimple\Container;
 
 class EwalletMessagingContainer extends Container
@@ -24,7 +23,7 @@ class EwalletMessagingContainer extends Container
         parent::__construct($arguments);
         $this->register(new DoctrineServiceProvider());
         $this->register(new TwigServiceProvider());
-        $this->register(new HexagonalServiceProvider());
+        $this->register(new MessagingServiceProvider());
         $this->register(new EwalletMessagingServiceProvider());
     }
 }
