@@ -4,23 +4,27 @@
  *
  * This source file is subject to the license that is bundled with this package in the file LICENSE.
  */
+
 namespace Ewallet\Pimple\ServiceProviders;
 
-use Hexagonal\RabbitMq\{AmqpMessageProducer, AmqpMessageConsumer, ChannelConfiguration};
-use Hexagonal\DomainEvents\StoredEvent;
-use Hexagonal\Messaging\{MessagePublisher, PublishedMessage};
+use Application\Messaging\MessagePublisher;
+use Application\Messaging\MessageTracker;
+use Doctrine\ORM\EntityManagerInterface;
+use Ports\Doctrine\Application\DomainEvents\EventStoreRepository;
+use Ports\Doctrine\Application\Messaging\MessageTrackerRepository;
 use PhpAmqpLib\Connection\AMQPStreamConnection;
-use Pimple\{Container, ServiceProviderInterface};
+use Pimple\Container;
+use Pimple\ServiceProviderInterface;
+use Ports\RabbitMq\Application\Messaging\AmqpMessageConsumer;
+use Ports\RabbitMq\Application\Messaging\AmqpMessageProducer;
+use Ports\RabbitMq\Application\Messaging\ChannelConfiguration;
 
-class HexagonalServiceProvider implements ServiceProviderInterface
+class MessagingServiceProvider implements ServiceProviderInterface
 {
     public function register(Container $container)
     {
-        $container['hexagonal.event_store_repository'] = function () use ($container) {
-            return $container['doctrine.em']->getRepository(StoredEvent::class);
-        };
-        $container['hexagonal.message_tracker_repository'] = function () use ($container) {
-            return $container['doctrine.em']->getRepository(PublishedMessage::class);
+        $container[MessageTracker::class] = function () use ($container) {
+            return new MessageTrackerRepository($container[EntityManagerInterface::class]);
         };
         $container['hexagonal.amqp_connection'] = function () use ($container) {
             return new AMQPStreamConnection(
@@ -47,8 +51,8 @@ class HexagonalServiceProvider implements ServiceProviderInterface
         };
         $container['hexagonal.messages_publisher'] = function () use ($container) {
             return new MessagePublisher(
-                $container['hexagonal.event_store_repository'],
-                $container['hexagonal.message_tracker_repository'],
+                $container[EventStoreRepository::class],
+                $container[MessageTracker::class],
                 $container['hexagonal.messages_producer']
             );
         };
