@@ -4,13 +4,17 @@
  *
  * This source file is subject to the license that is bundled with this package in the file LICENSE.
  */
+
 namespace Ewallet\SymfonyConsole\Commands;
 
+use Application\Messaging\MessageConsumer;
 use Closure;
-use Ewallet\ManageWallet\Notifications\{TransferFundsEmailNotifier, TransferFundsNotification};
-use Hexagonal\Messaging\MessageConsumer;
+use Ewallet\ManageWallet\Notifications\TransferFundsEmailNotifier;
+use Ewallet\ManageWallet\Notifications\TransferFundsNotification;
 use stdClass;
-use Symfony\Component\Console\{Command\Command, Input\InputInterface, Output\OutputInterface};
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
 
 class NotifyTransferByEmailCommand extends Command
 {
